discovery: move ServiceDiscoveryConfig
