+++ discovery/discoverer/config.go
@@ -11,7 +11,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package config
+package discoverer
+
+// TODO(abursavich): CLEANUP: this file's content will be deleted
 
 import (
 	"github.com/pkg/errors"
